Nonblocking worker is working, but the code is a mess.
Need to improve authbahn test reliability, re-enable a bunch of deleted unit
tests, and then start refactoring.

+++ src/server/thread_pool.zig
@@ -4,7 +4,7 @@ const Thread = std.Thread;
 const Allocator = std.mem.Allocator;
 
 pub const Opts = struct {
-	count: u32,
+	count: u16,
 	backlog: u32,
 	buffer_size: usize,
 };
